ucaip-notebooks/predictions/requirements.txt
joblib~=1.0
numpy~=1.19.2
scikit-learn~=0.24
tensorflow==2.4.1
google-cloud-storage>=1.26.0,<2.0.0dev
